+++ infra/core/orchestration/main.tf
@@ -22,6 +22,7 @@ locals {
   orchestration_key_pair_name                   = var.orchestration_key_pair_name
   orchestration_resource_prefix                 = var.orchestration_resource_prefix
   orchestration_source_security_group_ids       = var.orchestration_source_security_group_ids
+  pods_prefix                                    = var.pods_prefix
   private_subnets                               = var.private_subnets
   region                                        = var.region
   ssm_policy_for_instances_arn                  = var.ssm_policy_for_instances_arn
@@ -293,6 +294,7 @@ resource "aws_iam_policy" "orchestration_ec2_s3_access" {
         Resource = [
           "${local.data_bucket_arn}/*",
           "${local.infra_config_bucket_arn}/${local.orchestration_resource_prefix}/*",
+          "${local.infra_config_bucket_arn}/${local.pods_prefix}/*",
         ]
       },
       {
@@ -372,6 +374,24 @@ resource "aws_iam_policy" "orchestration_glue_policy" {
   })
 }
 
+resource "aws_iam_policy" "orchestration_polly_policy" {
+  name = "${local.environment}-${local.orchestration_resource_prefix}-polly-policy"
+  description = "Allow ec2 to call Polly"
+
+  policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Action = [
+          "polly:StartSpeechSynthesisTask",
+        ]
+        Effect = "Allow",
+        Resource = "*"
+      }
+    ]
+  })
+}
+
 resource "aws_iam_role_policy_attachment" "orchestration_role_ssm_policy_for_instances" {
   role       = aws_iam_role.orchestration_instance_role.name
   policy_arn = local.ssm_policy_for_instances_arn
@@ -400,6 +420,11 @@ resource "aws_iam_role_policy_attachment" "orchestration_role_secrets_policy" {
 resource "aws_iam_role_policy_attachment" "orchestration_role_glue_policy" {
   role       = aws_iam_role.orchestration_instance_role.name
   policy_arn = aws_iam_policy.orchestration_glue_policy.arn
+}
+
+resource "aws_iam_role_policy_attachment" "orchestration_role_polly_policy" {
+  role       = aws_iam_role.orchestration_instance_role.name
+  policy_arn = aws_iam_policy.orchestration_polly_policy.arn
 }
 
 # **********************************************************