+++ orchestration/airflow/dags/processing/parse_arxiv_summaries_dag.py
@@ -50,7 +50,7 @@ with DAG(
     SERVICE_NAME,
     catchup=False,
     default_args=DEFAULT_LOGGING_ARGS,
-    schedule_interval="@hourly",
+    schedule_interval="0 12 * * *",
     start_date=start_date,
     tags=["process", "arxiv"],
 ) as dag: