+++ orchestration/airflow/dags/shared/utils/constants.py
@@ -50,6 +50,49 @@ CS_CATEGORIES_INVERTED = {
     "NULL": "NULL",
 }
 
+CS_CATEGORIES = {
+    "AI": "AI",
+    "AR": "Hardware Architecture",
+    "CC": "Computational Complexity",
+    "CE": "Computational Engineering, Finance, and Science",
+    "CG": "Computational Geometry",
+    "CL": "NLP",
+    "CR": "Cryptography and Security",
+    "CY": "Computers and Society",
+    "CV": "Computer Vision",
+    "DB": "Databases",
+    "DC": "Distributed, Parallel, and Cluster Computing",
+    "DL": "Digital Libraries",
+    "DM": "Discrete Mathematics",
+    "DS": "Data Structures and Algorithms",
+    "ET": "Emerging Technologies",
+    "FL": "Formal Languages and Automata Theory",
+    "GL": "General Literature",
+    "GR": "Graphics",
+    "GT": "Computer Science and Game Theory",
+    "HC": "Human-Computer Interaction",
+    "IR": "Information Retrieval",
+    "IT": "Information Theory",
+    "LG": "Machine Learning",
+    "LO": "Logic in Computer Science",
+    "MA": "Multi-agent Systems",
+    "MM": "Multimedia",
+    "MS": "Mathematical Software",
+    "NA": "Numerical Analysis",
+    "NE": "Neural and Evolutionary Computing",
+    "NI": "Networking and Internet Architecture",
+    "OH": "Other Computer Science",
+    "OS": "Operating Systems",
+    "PF": "Performance",
+    "PL": "Programming Languages",
+    "SY": "Systems and Control",
+    "RO": "Robotics",
+    "SC": "Symbolic Computation",
+    "SD": "Sound",
+    "SE": "Software Engineering",
+    "SI": "Social and Information Networks",
+}
+
 # AWS
 AWS_REGION = "AWS_REGION"
 AWS_SECRETS_MANAGER = "secretsmanager"
@@ -133,6 +176,8 @@ PARSED_BY = "PARSED_BY"
 PARSES = "PARSES"
 PRIMARILY_CATEGORIZED_BY = "PRIMARILY_CATEGORIZED_BY"
 PROVIDES = "PROVIDES"
+PUBLISHED_IN = "PUBLISHED_IN"
+PUBLISHES = "PUBLISHES"
 RESEARCH_RECORD_DATE = "date"
 SUMMARIZED_BY = "SUMMARIZED_BY"
 SUMMARIZES = "SUMMARIZES"
@@ -154,6 +199,9 @@ ARXIV_API_MAX_RETRIES = "ARXIV_API_MAX_RETRIES"
 ARXIV_BASE_URL = "ARXIV_BASE_URL"
 ARXIV_INGESTION_DAY_SPAN = "ARXIV_INGESTION_DAY_SPAN"
 ARXIV_SETS = "ARXIV_SETS"
+CREATE_PODS_DAG = "create_pods_dag"
+CREATE_POD_TASK = "create_pod"
+CREATE_POD_TASK_VERSION = "CREATE_POD_TASK_VERSION"
 DATA_INGESTION_KEY_PREFIX = "DATA_INGESTION_KEY_PREFIX"
 DATA_ARXIV_SUMMARIES_INGESTION_COMPLETE_TOPIC = "data_arxiv_summaries_ingestion_complete"
 DISPATCH_LAMBDA_NAMES = "DISPATCH_LAMBDA_NAMES"
@@ -173,10 +221,14 @@ ORCHESTRATION_HOST_PRIVATE_IP = "ORCHESTRATION_HOST_PRIVATE_IP"
 PARSE_ARXIV_SUMMARIES_DAG = "parse_arxiv_summaries_dag"
 PARSE_SUMMARIES_TASK = "parse_summaries"
 PARSE_SUMMARIES_TASK_VERSION = "PARSE_SUMMARIES_TASK_VERSION"
+PARSE_SUMMARIES_TASK = "parse_summaries"
+PARSE_SUMMARIES_TASK_VERSION = "PARSE_SUMMARIES_TASK_VERSION"
 PERSIST_SUMMARIES_TASK = "persist_summaries"
 PERSIST_SUMMARIES_TASK_VERSION = "PERSIST_SUMMARIES_TASK_VERSION"
+PODS_PREFIX = "PODS_PREFIX"
 RAW_DATA_KEYS = "RAW_DATA_KEYS"
 RECORDS_PREFIX = "RECORDS_PREFIX"
+RETRIEVAL_ERRORS = "RETRIEVAL_ERRORS"
 SAVE_SUMMARIES_TO_DATALAKE_TASK = "save_summaries_to_datalake"
 SAVE_SUMMARIES_TO_DATALAKE_TASK_VERSION = "SAVE_SUMMARIES_TO_DATALAKE_TASK_VERSION"
 SCHEMA = "SCHEMA"