+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.6-alpha] - 2024-05-24
+
+### Added
+
+- Podcasts generated automatically daily
+- Podcast model
+
+### Removed
+
+- Podcasts audio generation manual trigger
+
 ## [0.3.5-alpha] - 2024-06-01
 
 ### Changed